Updated pinned versions for various dependencies. (#46)
In particular, tatami is updated, which in turn uses the new byteme
library. This fixes a bug with infinite looping on reading from a zlib
buffer, which was encountered when attempting to load the DICE
reference in loadLabelledReference.

We also switched to NIST's CMake-ified version of lib-h5wasm 
rather than using my ad hoc little fork.