Report identity vector instead of permutations for ScranMatrices. (#50)
This enables proper handling of row subsetting, which was not handled
properly by permutations (because there was no way to permute a
subset back to the full matrix). We can thus re-export subsetRows,
albeit with a few fixes to get it working again.

Similarly, we can now handle intersected cbind's in cbindWithRownames,
as this involved an implicit subset.