Register finalizer callbacks for deleting C++ class instances on GC. (#55)
This should reduce the chance of memory leaks in client applications when
developers forget to free the class instance (and thus the Wasm heap). Of
course, the finalizers may not be predictably called, so a manual free() is still
recommended for the most efficient release of unused memory. 

Also make all scran.js class members private to avoid people fiddling around
with them. Some still need to be exposed via getters to enable direct use
of the C++ instances by scran.js functions, but these getters are not documented,
so that should discourage their use by client applications.

Soft-deprecated safeFree and exported free() directly. Now that the deleting
is handled by the new release() function, there is no need to call delete()
directly, which removes the distinction between safeFree() and free().