Consistent handling of all buffer = null arguments.
If a function usually returns a WasmArray, then it will now return the input
'buffer' directly if it is supplied. All functions will now also use the
same 'local_buffer' pattern to avoid memory leaks when 'buffer = null'.

If a function usually returned a TypedArray, it will now return a TypedArray
view on the input 'buffer'. This is somewhat dangerous as the view may be
invalidated on the next Wasm allocation induced by the caller... but
nonetheless, a view is still more useful than a sudden 'undefined'.

Also added more tests for the usage of pre-specified buffers.