Eliminate row identities from all functions downstream of initialization. (#60)
This greatly simplifies operations on the ScranMatrix objects as we no longer
have to ensure that the row identities are always valid (e.g., for cbinding or
subsetting). Similarly, we are no longer obliged to copy the row identities
when creating new ScranMatrices, which cuts down on the potential for bugs.

Instead, the row identities are returned only in the initializeSparseMatrix*
functions when 'layered = true'. It is assumed that the caller uses them
immediately to permute the gene annotations to match the new order in the
layered matrix; after that, they can forget about the fact that there was any
reordering at all. Internally, we enforce this by wiping the identities from
the ScranMatrix so that there is no chance of affecting downstream code.

All downstream functionality that use the row identities inside a ScranMatrix
has been removed. cbind is now much stupider and just assumes that all inputs
have the same row order rather than trying to detect and correct differences.
Permutation functions no longer accept a ScranMatrix as input, but instead
require the separate identity vector. Delayed arithmetic on vectors no longer
provides an option to adjust for differences in row order. 

Deprecated functions based on row ordering have also been removed. I took the
opportunity to rename 'matchVectorToRowIdentities' to 'quickSliceArray', which
avoids any mention of row identities and is more general-purpose anyway.

Also updated docs and tests for this new paradigm.