Initialize sparse matrices from RDS objects. (#62)
In this manner, we can traverse a complex object inside an RDS file, identify
the matrix (or dg*Matrix), and use this to create a ScranMatrix.  The usual
layered= option is available, along with a consume= option to avoid copies by
moving precomputed CSC vectors into the output matrix when layered=false. (It
also allows in-place sorting of triplet vectors from a dgTMatrix.)