Added realizeFile to portably prepare files for reading.
This is most obviously useful for HDF5 readers, which have no choice but
to read from the file system. In such cases, the browser needs to read
from the VFS but Node.js can read from the actual FS. The new
realizeFile function abstracts away the decision of where to put the
file for input into the H5File constructor (amongst other things).