Overhauled the QC filtering for greater customization. (#68)
This propagates an update in the libscran C++ library. In particular, we now
have functions that compute (suggested) filter thresholds, which can be
customized by the user before applying them to generate a discard vector. This
is a breaking change as the previous classes are no longer available, given
that we no longer store the discard vector for each individual metric. 

In addition:
- Modality-specific functions are also more consistently named.
- Updated the README with the new functions.
- Run tests in serial to avoid timeout issues.