Prevent users from using unfilled fields in empty objects.
This is achieved by setting 'filled = false' in the empty*Results() functions.
If anyone attempts to call a getter while 'filled = false', an error is raised.
The flag is set to true if they use a setter or they create a memory view with
'fillable = true', at which point it is assumed that some setting was done.