Support scaling to unit variance in scoreFeatureSet.
In small sets, the score may be dominated by a very small number of features.
Whether this is good or bad is debatable - e.g., if only one feature is highly
variable, a large weight for that feature makes sense - but users should be
able to scale to negate any single-feature dominance if they feel it's wrong.