summary= argument to ScoreMarkersResults is now a string.
This avoids the need for cryptic integer encodings; rather, users can just pass
along summary:"mean" or similar, which makes for easier reading.

Also added support for computing median and maxima effect sizes. These are not
commonly used; the median does the same as the mean but requires more work,
while the maxima is only useful for looking at downregulation. Nonetheless, we
provide bindings to compute them and to fill them in empty objects.