Added general-purpose utility to convert arrays into R-style factors.
factorize() elaborates on the existing convertBlock function, with protection
against invalid entries (null, NaN) and support for Int32Array outputs. This
enables more general use, particularly to create the groups= for scoreMarkers.

convertBlock() now calls factorize() under the hood, with a slightly new
behavior in that it throws an error if invalid values are detected (as opposed
to letting them get silently filled in with who-knows-what). convertBlock()
itself is now also soft-deprecated as factorize() provides a superset of its
functionality and is more appropriately named for general use.

Bumped version and updated NEWS to get this published.