Generalized the block subsetting to work for all factors.
This now handles factors with both Int32Array and Int32WasmArray ids. It
also performs automatic dropping of unused levels, given that other
scran.js functions expect all levels to be used.

This takes advantage of the latest wasmarrays.js package, which handles
the subsetting of WasmArrays more easily. The subsetting logic was also
migrated to wasmarrays.js for more general use. This allows us to delete
a lot of the block subsetting code entirely.