Simplify the AggregateAcrossCellsResults interface.
Rather than passing back a ScranMatrix, we return the concatenated array
across all groups, and advanced users can use this to construct their
own ScranMatrix via the new createDenseMatrix function. This is a more
general solution that avoids us having to write a ScranMatrix converter
for every dense output; it is also more obvious to the user that the
ScranMatrix needs to have its memory lifetime carefully managed if it
holds a reference to the same buffer as the AggregateResults object.