Improved feature ID intersection in labelCells. (#73)
We now support for synonyms in the referenceFeatures, e.g., due to 1:many
mappings between gene ID types. If an array of synonyms is supplied, the first
synonym that matches a feature from the dataset is used. 

We also explicitly use first occurrence of duplicated IDs in all feature
processing. So, only the first occurrence of a feature ID within 'features' is
used. Similarly, if multiple features in 'referenceFeatures' match to the same
entry of 'features', only the first matching entry is used.