Allow larger cache sizes to handle awkward HDF5 dense chunks. (#78)
This refers to the situation where someone has provided a dense matrix in a
HDF5 file and the chunks are slightly too large, such that we cannot cache a
full column/row's worth of chunks. By increasing the cache size beyond the
default, we can cache properly and avoid repeated reads on every request.