Update to the latest versions of C++ libraries. (#81)
This is a major version change that incorporates the latest libscran and tatami
libraries. We eliminate the need to manually reorder when using layered
matrices; we also introduce a consistent weighting scheme across all
block-averaging functions, and we add a new block-handling mode for the PCA.
Many functions were also renamed to conform to camelCase/PascalCase.
Check out the NEWS file for more details.

We removed some legacy stuff such as the doxygen mark-up in the C++ code (who
reads that anyway?), some threading-related macros, and constructors for 
the result objects that were used in the defunct bakana state loaders.

The Dockerfile has been moved to its own separate repository so that the docker
image can be updated independently of the scran.js source code. This avoids 
problems where a PR needs to modify the Dockerfile and also contains changes
that depend on the same modification, but the docker image can't be updated
until it gets merged into the master branch!