Only create extractors on first row/col request.
This avoids the overhead of extractor creation when the NumericMatrix is not
going to be queried from JS, e.g., because it's an intermediate. The downside
is that the row/col getters are no longer thread-safe, but they can only
be called from JS and that should be single-threaded anyway.