Switch back to internal builds of igraph and HDF5. (#89)
This avoids headaches from guessing ABI compatibility across different
Emscripten versions and build flags. With suitable caching, we can avoid
achieve prebuilt binaries but with more suitable integration with our own CI.

We also abandon the dedicated scran.js Docker image in favor of the generic
emsdk + CMake image. This reduces the complexity of the CI and avoids stale
object files in the CMake build. We use more caching to save the Wasm binary so
that it doesn't need to be rebuilt if the C++ code hasn't changed.

Also got rid of preamble.js as it's no longer needed in the new toolchain.