Upgrade to the latest version of all C++ libraries. (#100)
This doesn't change the output for any of the functions, except for MNN
correction where the algorithm changed in v3, and for the graph clustering
functions where igraph v1.0.0 might have introduced some changes.

Introduced a JsFakeInt typedef to make it clear where we're expecting
an integer from a Javascript Number and vice versa. Also added utilities
to safely coerce to/from actual integer types.

Updated HDF5 to 2.0.0 to support new files created in the new format.