Replace all uintptr_t's with JsFakeInts for consistency.
All other integer inputs are accepted as JsFakeInts and are safely cast within
the function body, so we do the same with the uintptr_t's (which are still
represented as Javascript numbers from _malloc and in the $$.ptrs).

Also fixed a few other things while we're at it, such as making sure all
integers are passed as JsFakeInts, and namespacing the fixed-width types.
In particular, cleaned up the NeighborResults constructor.